+++ tests/faces/test_bsplineface3d.py
@@ -1,8 +1,17 @@
 import unittest
-from volmdlr import surfaces, wires, faces
+
+import volmdlr
+from volmdlr import edges, surfaces, wires, faces
+from volmdlr.models.bspline_surfaces import bspline_surface_1
 
 
-class MyTestCase(unittest.TestCase):
+class TestBSplineFace3D(unittest.TestCase):
+    bspline_face = faces.BSplineFace3D.from_surface_rectangular_cut(bspline_surface_1, 0, 1, 0, 1)
+
+    def test_is_linesegment_crossing(self):
+        linesegment = edges.LineSegment3D(volmdlr.Point3D(4, 0, 0), volmdlr.Point3D(4, 2, 2))
+        self.assertTrue(self.bspline_face.is_linesegment_crossing(linesegment=linesegment))
+
     def test_from_contours3d(self):
         surface = surfaces.BSplineSurface3D.load_from_file(
             "faces/objects_bspline_test/bspline_surface_openned_contour.json")